RF Merge Folder13Test into FolderTest
